Add console information to learning section
Added some information about the console feature in the learning section
which may be useful to some, added a section that describes why you'll
have issues if there's no 2d window open. Will link to this url in
the exception message the user sees.

Also added Jekyll to the readme as it's needed to run the site locally.